Added a stack class
Added a constructor to the stack class.


Fixed stack class bug.

The bug causes the stack to return an IndexOutofBoundsException from the underlying linked list.

This is caused by a bug in the extendList() function in the Queue class.

The Queue extends the list by adding five new elements to the underlying pseudo-array every time the Queue needs more space. It must then copy the elements from the old pseudo-array to the newly-created array.

However, the extendList() function copies the elements of the old array from index 0 to index `size`. `size` is the specified number of elements to add to the pseudo-array. The bug occurs for Queues that have an initial size less than `size`. extendList() attempts to access a non-existent element from the old pseudo-array.

The fix was to use the number of elements present in the old pseudo-array as the upper limit for the counter for extendList() when deep copying elements over to the new array.